[TASK] Adjust unit tests
